input verification v2 + user verification v2 + swagger overhaul v1

+++ src/controller/blog.controller.ts
@@ -37,16 +37,20 @@ interface likeCreate {
 export const createBlog: Handler = async (req, res) => {
     try {
         const token = req.cookies.access_token
+        if (!token) throw 'No cookie/token'
         const decoded = jwt.verify(token, process.env.SECRET_JWT_KEY!) as JwtPayload
         const reqCreate: createInterface = { title: req.body.title, content: req.body.content || '', authorId: decoded.id, popularity: '0%' }
         if (typeof reqCreate.title != 'string') throw 'Invalid title'
         if (typeof reqCreate.content != 'string') throw 'Invalid content'
+        if (reqCreate.authorId == null) throw 'No Author ID'
         const newBlog = await blogModel.createBlog(reqCreate)
         if (!newBlog) throw 'Not created'
         res.status(201).send(newBlog)
     } catch (err) {
+        if (err == 'No cookie/token') res.status(400).send('There is no cookie or token')
         if (err == 'Invalid title') res.status(400).send('This title is invalid')
         if (err == 'Invalid content') res.status(400).send('This content is invalid')
+        if (err == 'No Author ID') res.status(400).send('No Author ID is being sent')
         if (err == 'Not created') res.status(422).send('This blog could not be created')
     }
 }
@@ -54,14 +58,20 @@ export const createBlog: Handler = async (req, res) => {
 export const likeBlog: Handler = async (req, res) => {
     try {
         const token = req.cookies.access_token
+        if (!token) throw 'No cookie/token'
         const decoded = jwt.verify(token, process.env.SECRET_JWT_KEY!) as JwtPayload
         const userFind: authorFind = { authorId: decoded.id }
+        if (userFind.authorId == null) throw 'No Author ID ver'
         const user = await blogModel.getAuthor(userFind)
         if (!user) throw 'No user'
         const blogFind: blogFind = { id: req.params.id }
+        if (blogFind.id == null) throw 'No ID ver'
         const specificBlog = await blogModel.getBlog(blogFind)
         if (!specificBlog) throw 'No blog'
+        if (specificBlog.authorId == userFind.authorId) throw 'Same user'
         const reqFind: likeCreate = { blogId: +req.params.id, userId: decoded.id }
+        if (reqFind.blogId == null) throw 'No ID like'
+        if (reqFind.userId == null) throw 'No Author ID like'
         const findLike = await blogModel.getLike(reqFind)
         if (findLike) throw 'Existing like'
         const createLike = await blogModel.likeBlog(reqFind)
@@ -72,8 +82,14 @@ export const likeBlog: Handler = async (req, res) => {
         if (!popUpdate) throw 'Popularity not calculated'
         res.status(200).send('The blog was liked correctly')
     } catch (err) {
+        if (err == 'No cookie/token') res.status(400).send('There is no cookie or token')
+        if (err == 'No Author ID ver') res.status(400).send('No Author ID is being sent on author verification')
         if (err == 'No user') res.status(404).send('The user does not exist')
+        if (err == 'No ID ver') res.status(400).send('No ID is being sent on blog verification')
         if (err == 'No blog') res.status(404).send('The blog does not exist')
+        if (err == 'Same user') res.status(400).send('A user cannot like their own post')
+        if (err == 'No ID like') res.status(400).send('No ID is being sent on like request')
+        if (err == 'No Author ID like') res.status(400).send('No Author ID is being sent on like request')
         if (err == 'Existing like') res.status(400).send('The like already exists')
         if (err == 'Unlikable') res.status(422).send('The like could not be created')
         if (err == 'Counter not increased') res.status(400).send('The like could not be counted properly')
@@ -84,14 +100,19 @@ export const likeBlog: Handler = async (req, res) => {
 export const dislikeBlog: Handler = async (req, res) => {
     try {
         const token = req.cookies.access_token
+        if (!token) throw 'No cookie/token'
         const decoded = jwt.verify(token, process.env.SECRET_JWT_KEY!) as JwtPayload
         const userFind: authorFind = { authorId: decoded.id }
+        if (userFind.authorId == null) throw 'No Author ID ver'
         const user = await blogModel.getAuthor(userFind)
         if (!user) throw 'No user'
         const blogFind: blogFind = { id: req.params.id }
+        if (blogFind.id == null) throw 'No ID ver'
         const specificBlog = await blogModel.getBlog(blogFind)
         if (!specificBlog) throw 'No blog'
         const reqFind: likeCreate = { blogId: +req.params.id, userId: decoded.id }
+        if (reqFind.blogId == null) throw 'No ID like'
+        if (reqFind.userId == null) throw 'No Author ID like'
         const findLike = await blogModel.getLike(reqFind)
         if (!findLike) throw 'Non-existing like'
         const createDislike = await blogModel.dislikeBlog(reqFind)
@@ -102,8 +123,13 @@ export const dislikeBlog: Handler = async (req, res) => {
         if (!popUpdate) throw 'Popularity not calculated'
         res.status(204).send('The blog was disliked correctly')
     } catch (err) {
+        if (err == 'No cookie/token') res.status(400).send('There is no cookie or token')
+        if (err == 'No Author ID ver') res.status(400).send('No Author ID is being sent on author verification')
         if (err == 'No user') res.status(404).send('The user does not exist')
+        if (err == 'No ID ver') res.status(400).send('No ID is being sent on blog verification')
         if (err == 'No blog') res.status(404).send('The blog does not exist')
+        if (err == 'No ID like') res.status(400).send('No ID is being sent on like request')
+        if (err == 'No Author ID like') res.status(400).send('No Author ID is being sent on like request')
         if (err == 'Non-existing like') res.status(400).send('The like does not exist')
         if (err == 'Undislikable') res.status(422).send('The like could not be reverted')
         if (err == 'Counter not decreased') res.status(400).send('The dislike could not be counted properly')
@@ -124,10 +150,12 @@ export const getBlogs: Handler = async (req, res) => {
 export const getBlog: Handler = async (req, res) => {
     try {
         const reqFind: blogFind = { id: req.params.id }
+        if (reqFind.id == null) throw 'No ID'
         const specificBlog = await blogModel.getBlog(reqFind)
         if (!specificBlog) throw 'Empty'
         res.status(200).send(specificBlog)
     } catch (err) {
+        if (err == 'No ID') res.status(400).send('No ID is being sent')
         if (err == 'Empty') res.status(404).send('This blog could not be found')
     }
 }
@@ -135,14 +163,18 @@ export const getBlog: Handler = async (req, res) => {
 export const getBlogsByAuthor: Handler = async (req, res) => {
     try {
         const token = req.cookies.access_token
+        if (!token) throw 'No cookie/token'
         const decoded = jwt.verify(token, process.env.SECRET_JWT_KEY!) as JwtPayload
         const reqFind: authorFind = { authorId: decoded.id }
+        if (reqFind.authorId == null) throw 'No Author ID'
         const user = await blogModel.getAuthor(reqFind)
         if (!user) throw 'No user'
         const authorBlogs = await blogModel.getBlogsByAuthor(reqFind)
         if (!authorBlogs) throw 'Empty'
         res.status(200).send(authorBlogs)
     } catch (err) {
+        if (err == 'No cookie/token') res.status(400).send('There is no cookie or token')
+        if (err == 'No Author ID') res.status(400).send('No Author ID is being sent')
         if (err == 'No user') res.status(404).send('This user does not exist')
         else if (err == 'Empty') res.status(404).send('No blogs from this author could be found')
     }
@@ -150,22 +182,50 @@ export const getBlogsByAuthor: Handler = async (req, res) => {
 
 export const deleteBlog: Handler = async (req, res) => {
     try {
+        const token = req.cookies.access_token
+        if (!token) throw 'No cookie/token'
+        const decoded = jwt.verify(token, process.env.SECRET_JWT_KEY!) as JwtPayload
+        const userFind: authorFind = { authorId: decoded.id }
+        if (userFind.authorId == null) throw 'No Author ID'
         const reqDelete: blogFind = { id: req.params.id }
+        if (reqDelete.id == null) throw 'No ID'
+        const specificBlog = await blogModel.getBlog(reqDelete)
+        if (!specificBlog) throw 'No blog'
+        if (specificBlog.authorId != decoded.id) throw 'Unauthorized user'
         const blogDelete = await blogModel.deleteBlog(reqDelete)
         if (!blogDelete) throw 'Empty'
         res.status(204).send('The blog was deleted correctly')
     } catch (err) {
+        if (err == 'No cookie/token') res.status(400).send('There is no cookie or token')
+        if (err == 'No Author ID') res.status(400).send('No Author ID is being sent')
+        if (err == 'No ID') res.status(400).send('No ID is being sent')
+        if (err == 'No blog') res.status(404).send('The blog does not exist')
+        if (err == 'Unauthorized user') res.status(401).send('A user can only delete their own posts')
         if (err == 'Empty') res.status(400).send('This blog could not be deleted')
     }
 }
 
 export const recoverBlog: Handler = async (req, res) => {
     try {
+        const token = req.cookies.access_token
+        if (!token) throw 'No cookie/token'
+        const decoded = jwt.verify(token, process.env.SECRET_JWT_KEY!) as JwtPayload
+        const userFind: authorFind = { authorId: decoded.id }
+        if (userFind.authorId == null) throw 'No Author ID'
         const reqRecover: blogFind = { id: req.params.id }
+        if (reqRecover.id == null) throw 'No ID'
+        const specificBlog = await blogModel.getBlog(reqRecover)
+        if (!specificBlog) throw 'No blog'
+        if (specificBlog.authorId != decoded.id) throw 'Unauthorized user'
         const blogRecover = await blogModel.recoverBlog(reqRecover)
         if (!blogRecover) throw 'Empty'
         res.status(200).send(blogRecover)
     } catch (err) {
+        if (err == 'No cookie/token') res.status(400).send('There is no cookie or token')
+        if (err == 'No Author ID') res.status(400).send('No Author ID is being sent')
+        if (err == 'No ID') res.status(400).send('No ID is being sent')
+        if (err == 'No blog') res.status(404).send('The blog does not exist')
+        if (err == 'Unauthorized user') res.status(401).send('A user can only recover their own posts')
         if (err == 'Empty') res.status(400).send('This blog could not be recovered')
     }
 }
@@ -173,6 +233,7 @@ export const recoverBlog: Handler = async (req, res) => {
 export const updateBlog: Handler = async (req, res) => {
     try {
         const reqFind: blogFind = { id: req.params.id }
+        if (reqFind.id == null) throw 'No ID'
         const blog = await blogModel.getBlog(reqFind)
         if (!blog) throw 'No blog'
         const reqUpdate: blogUpdate = { title: req.body.title, content: req.body.content }
@@ -182,6 +243,7 @@ export const updateBlog: Handler = async (req, res) => {
         if (!blogUpdate) throw 'Empty'
         res.status(200).send(blogUpdate)
     } catch (err) {
+        if (err == 'No ID') res.status(400).send('No ID is being sent')
         if (err == 'No blog') res.status(404).send('This blog could not be found')
         if (err == 'Invalid title') res.status(400).send('This title is invalid')
         if (err == 'Invalid content') res.status(400).send('This content is invalid')
@@ -192,16 +254,20 @@ export const updateBlog: Handler = async (req, res) => {
 export const actuallyDeleteBlog: Handler = async (req, res) => {
     try {
         const token = req.cookies.access_token
+        if (!token) throw 'No cookie/token'
         const decoded = jwt.verify(token, process.env.SECRET_JWT_KEY!) as JwtPayload
         if (decoded.role == 'simpleUser') throw 'Unauthorized'
         const reqDelete: blogFind = { id: req.params.id }
+        if (reqDelete.id == null) throw 'No ID'
         const likesFinalDelete = await blogModel.deleteLikes(reqDelete)
         if (!likesFinalDelete) throw 'Likes undeletable'
         const blogFinalDelete = await blogModel.actuallyDeleteBlog(reqDelete)
         if (!blogFinalDelete) throw 'Undeletable'
         res.status(204).send('The blog was completely deleted correctly')
     } catch (err) {
+        if (err == 'No cookie/token') res.status(400).send('There is no cookie or token')
         if (err == 'Unauthorized') res.status(401).send('The user does not have the necessary access level')
+        if (err == 'No ID') res.status(400).send('No ID is being sent')
         if (err == 'Likes undeletable') res.status(400).send('The likes from this blog could not be completely deleted')
         if (err == 'Undeletable') res.status(400).send('This blog could not be completely deleted')
     }
@@ -210,24 +276,16 @@ export const actuallyDeleteBlog: Handler = async (req, res) => {
 export const popularityScore: Handler = async (req, res) => {
     try {
         const reqFind: blogFind = { id: req.params.id }
+        if (reqFind.id == null) throw 'No ID'
         const popularity = await blogModel.popularityScore(reqFind)
         if (!popularity) throw 'No popularity score'
         res.status(200).send('Popularity updated')
     } catch (err) {
+        if (err == 'No ID') res.status(400).send('No ID is being sent')
         if (err == 'No popularity score') res.status(400).send("This blog's popularity score could not be updated")
     }
 }
 // - - - - - - - -
-export const increaseCounter: Handler = async (req, res) => {
-    try {
-        const reqFind: blogFind = { id: req.params.id }
-        const blogCounter = await blogModel.increaseCounter(reqFind)
-        if (!blogCounter) throw 'Empty'
-        res.status(200).send(blogCounter)
-    } catch (err) {
-        if (err == 'Empty') res.status(400).send("This blog's like counter could not be increased")
-    }
-}
 
 export const deleteAllBlogs: Handler = async (req, res) => {
     try {
@@ -246,26 +304,5 @@ export const deleteAllLikes: Handler = async (req, res) => {
         res.status(204).send('The likes were deleted correctly')
     } catch (err) {
         if (err == 'Undeletable') res.status(400).send('The likes could not be deleted')
-    }
-}
-
-export const testing: Handler = async (req, res) => {
-    try {
-        const likes = await blogModel.getLikes()
-        if (!likes) throw 'Empty'
-        res.status(200).send(likes)
-    } catch (err) {
-        if (err == 'Empty') res.status(400).send('The testing did not work')
-    }
-}
-
-export const testingTwo: Handler = async (req, res) => {
-    try {
-        const reqFind: likeCreate = { blogId: +req.body.id, userId: +req.body.userId }
-        const likes = await blogModel.dislikeBlog(reqFind)
-        if (!likes) throw 'Empty'
-        res.status(200).send(likes)
-    } catch (err) {
-        if (err == 'Empty') res.status(400).send('The testingTwo did not work')
     }
 }
